Disables Whole Program Optimization
Whole Program Optimization is problematic for static-libs.
Any project trying to utilize syscommon must be utilizing the same
compiler version at the build libraries, which is a maintenance burden
nobody want.

As such, it is generally a good idea to ensure it is disabled for
libraries, especially those not included directly in a project.